OTV: Move states to own files (#412)
This change moves each state into it's own file to make the states more manageable and easier to find.
