+++ cmd/oceantv/broadcast_states.go
@@ -153,381 +153,6 @@ func (s *liveStateFields) lastChatMsg() time.Time         { return s.LastChatMsg
 func (s *liveStateFields) setLastStatusCheck(t time.Time) { s.LastStatusCheck = t }
 func (s *liveStateFields) setLastChatMsg(t time.Time)     { s.LastChatMsg = t }
 
-type vidforwardPermanentStarting struct {
-	stateFields
-	stateWithTimeoutFields
-}
-
-func newVidforwardPermanentStarting(ctx *broadcastContext) *vidforwardPermanentStarting {
-	return &vidforwardPermanentStarting{stateWithTimeoutFields: newStateWithTimeoutFields(ctx)}
-}
-
-func (s *vidforwardPermanentStarting) enter() {
-	s.LastEntered = time.Now()
-
-	// Use a copy of the config so that we can adjust the end date to +1 year
-	// without affecting the original config.
-	cfg := *s.cfg
-	cfg.End = cfg.End.AddDate(1, 0, 0)
-
-	if !try(s.man.SetupSecondary(context.Background(), s.cfg, s.store), "could not setup secondary broadcast", s.log) {
-		s.bus.publish(startFailedEvent{})
-		return
-	}
-
-	// We pass this to createAndStart so that it's run after broadcast creation, therefore
-	// vidforward gets up to date RTMP endpoint information.
-	onBroadcastCreation := func() error {
-		err := s.fwd.Stream(&cfg)
-		if err != nil {
-			return fmt.Errorf("could not set vidforward mode to stream: %w", err)
-		}
-		return nil
-	}
-
-	createBroadcastAndRequestHardware(
-		s.broadcastContext,
-		&cfg,
-		onBroadcastCreation,
-	)
-}
-
-type vidforwardPermanentLive struct {
-	stateFields
-	liveStateFields
-}
-
-func newVidforwardPermanentLive() *vidforwardPermanentLive { return &vidforwardPermanentLive{} }
-
-type vidforwardPermanentTransitionLiveToSlate struct {
-	stateFields
-	stateWithTimeoutFields
-	stateWithHealthFields
-	HardwareStopped bool
-}
-
-func newVidforwardPermanentTransitionLiveToSlate(ctx *broadcastContext) *vidforwardPermanentTransitionLiveToSlate {
-	return &vidforwardPermanentTransitionLiveToSlate{stateWithTimeoutFields: newStateWithTimeoutFields(ctx)}
-}
-
-func (s *vidforwardPermanentTransitionLiveToSlate) enter() {
-	s.LastEntered = time.Now()
-
-	s.bus.publish(hardwareStopRequestEvent{})
-	try(s.fwd.Slate(s.cfg), "could not set vidforward mode to slate", s.log)
-}
-func (s *vidforwardPermanentTransitionLiveToSlate) isHardwareStopped() bool {
-	return s.cfg.HardwareState == hardwareStateToString(&hardwareOff{})
-}
-
-type vidforwardPermanentTransitionSlateToLive struct {
-	stateFields
-	HardwareStarted bool
-	stateWithTimeoutFields
-	stateWithHealthFields
-}
-
-func newVidforwardPermanentTransitionSlateToLive(ctx *broadcastContext) *vidforwardPermanentTransitionSlateToLive {
-	return &vidforwardPermanentTransitionSlateToLive{stateWithTimeoutFields: newStateWithTimeoutFields(ctx)}
-}
-func (s *vidforwardPermanentTransitionSlateToLive) enter() {
-	s.LastEntered = time.Now()
-	s.bus.publish(hardwareStartRequestEvent{})
-	try(s.fwd.Stream(s.cfg), "could not set vidforward mode to stream", s.log)
-}
-func (s *vidforwardPermanentTransitionSlateToLive) isHardwareStarted() bool {
-	return s.cfg.HardwareState == hardwareStateToString(&hardwareOn{})
-}
-
-type vidforwardPermanentLiveUnhealthy struct {
-	stateFields
-	*broadcastContext `json: "-"`
-	LastResetAttempt  time.Time
-	Attempts          int
-	liveStateFields
-}
-
-func newVidforwardPermanentLiveUnhealthy(ctx *broadcastContext) *vidforwardPermanentLiveUnhealthy {
-	return &vidforwardPermanentLiveUnhealthy{broadcastContext: ctx}
-}
-func (s *vidforwardPermanentLiveUnhealthy) fix() {
-	const resetInterval = 5 * time.Minute
-	if time.Since(s.LastResetAttempt) <= resetInterval {
-		return
-	}
-
-	s.Attempts++
-
-	var (
-		e   event
-		msg string
-	)
-
-	const maxAttempts = 3
-	if s.Attempts > maxAttempts {
-		msg = "failed to fix permanent broadcast, transitioning to slate (attempts: %d, max attempts: %d)"
-		e = fixFailureEvent{}
-	} else {
-		msg = "attempting to fix permanent broadcast by hardware restart and forward stream re-request (attempts: %d, max attempts: %d)"
-		try(s.fwd.Stream(s.cfg), "could not set vidforward mode to stream", s.log)
-		e = hardwareResetRequestEvent{}
-	}
-
-	s.logAndNotify(broadcastGeneric, msg, s.Attempts, maxAttempts)
-	s.bus.publish(e)
-	s.LastResetAttempt = time.Now()
-}
-
-type vidforwardPermanentFailure struct {
-	stateFields
-	*broadcastContext `json: "-"`
-}
-
-func newVidforwardPermanentFailure(ctx *broadcastContext) *vidforwardPermanentFailure {
-	return &vidforwardPermanentFailure{stateFields{}, ctx}
-}
-func (s *vidforwardPermanentFailure) enter() { s.requestSlate() }
-func (s *vidforwardPermanentFailure) fix()   { s.requestSlate() }
-func (s *vidforwardPermanentFailure) requestSlate() {
-	s.bus.publish(hardwareStopRequestEvent{})
-	try(s.fwd.Slate(s.cfg), "could not set vidforward mode to slate", s.log)
-}
-
-type vidforwardPermanentSlate struct{ stateFields }
-
-func newVidforwardPermanentSlate() *vidforwardPermanentSlate { return &vidforwardPermanentSlate{} }
-
-type vidforwardPermanentVoltageRecoverySlate struct {
-	stateFields
-	stateWithTimeoutFields
-}
-
-func newVidforwardPermanentVoltageRecoverySlate(ctx *broadcastContext) *vidforwardPermanentVoltageRecoverySlate {
-	return &vidforwardPermanentVoltageRecoverySlate{
-		stateWithTimeoutFields: newStateWithTimeoutFieldsWithTimeout(
-			ctx,
-			time.Duration(sanatisedVoltageRecoveryTimeout(ctx))*time.Hour,
-		),
-	}
-}
-func (s *vidforwardPermanentVoltageRecoverySlate) enter() {
-	s.LastEntered = time.Now()
-	s.requestSlate()
-}
-func (s *vidforwardPermanentVoltageRecoverySlate) fix() { s.requestSlate() }
-func (s *vidforwardPermanentVoltageRecoverySlate) requestSlate() {
-	try(s.fwd.Slate(s.cfg, WithType(LowVoltage)), "could not set vidforward mode to low voltage slate", s.log)
-}
-
-type vidforwardPermanentSlateUnhealthy struct {
-	stateFields
-	*broadcastContext `json: "-"`
-	LastResetAttempt  time.Time
-}
-
-func newVidforwardPermanentSlateUnhealthy(ctx *broadcastContext) *vidforwardPermanentSlateUnhealthy {
-	return &vidforwardPermanentSlateUnhealthy{stateFields{}, ctx, time.Now()}
-}
-func (s *vidforwardPermanentSlateUnhealthy) fix() {
-	const resetInterval = 5 * time.Minute
-	if time.Since(s.LastResetAttempt) > resetInterval {
-		s.logAndNotify(broadcastForwarder, "slate is unhealthy, requesting vidforward reconfiguration")
-		try(s.fwd.Slate(s.cfg), "could not set vidforward mode to slate", s.log)
-		s.LastResetAttempt = time.Now()
-	}
-}
-
-type vidforwardPermanentIdle struct {
-	stateFields
-	*broadcastContext `json: "-"`
-}
-
-func newVidforwardPermanentIdle(ctx *broadcastContext) *vidforwardPermanentIdle {
-	return &vidforwardPermanentIdle{stateFields{}, ctx}
-}
-func (s *vidforwardPermanentIdle) enter() {
-	s.bus.publish(hardwareStopRequestEvent{})
-}
-
-type vidforwardSecondaryLive struct {
-	stateFields
-	*broadcastContext `json: "-"`
-	liveStateFields
-}
-
-func newVidforwardSecondaryLive(ctx *broadcastContext) *vidforwardSecondaryLive {
-	return &vidforwardSecondaryLive{broadcastContext: ctx}
-}
-func (s *vidforwardSecondaryLive) exit() {
-	err := s.man.StopBroadcast(context.Background(), s.cfg, s.store, s.svc)
-	if err != nil {
-		s.log("could not stop broadcast on exit: %v", err)
-		return
-	}
-	s.bus.publish(finishedEvent{})
-}
-
-type vidforwardSecondaryLiveUnhealthy struct {
-	stateFields
-	liveStateFields
-}
-
-func newVidforwardSecondaryLiveUnhealthy() *vidforwardSecondaryLiveUnhealthy {
-	return &vidforwardSecondaryLiveUnhealthy{}
-}
-
-type vidforwardSecondaryStarting struct {
-	stateFields
-	stateWithTimeoutFields
-}
-
-func newVidforwardSecondaryStarting(ctx *broadcastContext) *vidforwardSecondaryStarting {
-	return &vidforwardSecondaryStarting{stateWithTimeoutFields: newStateWithTimeoutFields(ctx)}
-}
-
-func (s *vidforwardSecondaryStarting) enter() {
-	s.LastEntered = time.Now()
-	// We pass this to createBroadcastAndRequestHardware so that it's run after
-	// broadcast creation, therefore vidforward gets up to date RTMP endpoint
-	// information.
-	onBroadcastCreation := func() error {
-		err := s.fwd.Stream(s.cfg)
-		if err != nil {
-			return fmt.Errorf("could not set vidforward mode to stream: %w", err)
-		}
-		return nil
-	}
-	createBroadcastAndRequestHardware(
-		s.broadcastContext,
-		s.cfg,
-		onBroadcastCreation,
-	)
-}
-
-type vidforwardSecondaryIdle struct {
-	stateFields
-	*broadcastContext `json: "-"`
-}
-
-func newVidforwardSecondaryIdle(ctx *broadcastContext) *vidforwardSecondaryIdle {
-	return &vidforwardSecondaryIdle{broadcastContext: ctx}
-}
-func (s *vidforwardSecondaryIdle) enter() {
-	s.bus.publish(hardwareStopRequestEvent{})
-}
-
-type directLive struct {
-	*broadcastContext `json: "-"`
-	stateFields
-	liveStateFields
-}
-
-func newDirectLive(ctx *broadcastContext) *directLive {
-	return &directLive{broadcastContext: ctx}
-}
-
-type directLiveUnhealthy struct {
-	*broadcastContext `json: "-"`
-	LastResetAttempt  time.Time
-	Attempts          int
-	stateFields
-	liveStateFields
-}
-
-func newDirectLiveUnhealthy(ctx *broadcastContext) *directLiveUnhealthy {
-	return &directLiveUnhealthy{broadcastContext: ctx}
-}
-func (s *directLiveUnhealthy) fix() {
-	const directLiveFixTimeout = 8 * time.Minute
-	if time.Since(s.LastResetAttempt) <= directLiveFixTimeout {
-		return
-	}
-
-	var e event
-	const maxAttempts = 3
-	if s.Attempts >= maxAttempts {
-		e = fixFailureEvent{fmt.Errorf("failed to fix broadcast (attempts: %d, max attempts: %d)", s.Attempts, maxAttempts)}
-	} else {
-		s.logAndNotify(broadcastHardware, "attempting to fix broadcast by hardware restart request (attempts: %d, max attempts: %d)", s.Attempts, maxAttempts)
-		s.Attempts++
-		e = hardwareResetRequestEvent{}
-	}
-
-	s.LastResetAttempt = time.Now()
-	s.bus.publish(e)
-}
-
-type directStarting struct {
-	stateFields
-	stateWithTimeoutFields
-}
-
-func newDirectStarting(ctx *broadcastContext) *directStarting {
-	return &directStarting{stateWithTimeoutFields: newStateWithTimeoutFields(ctx)}
-}
-func (s *directStarting) enter() {
-	s.LastEntered = time.Now()
-	createBroadcastAndRequestHardware(s.broadcastContext, s.cfg, nil)
-}
-
-type directIdle struct {
-	stateFields
-	*broadcastContext `json: "-"`
-}
-
-func newDirectIdle(ctx *broadcastContext) *directIdle { return &directIdle{broadcastContext: ctx} }
-func (s *directIdle) enter() {
-	err := s.man.StopBroadcast(context.Background(), s.cfg, s.store, s.svc)
-	if err != nil {
-		s.log("could not stop broadcast on entry: %v", err)
-	} else {
-		s.bus.publish(finishedEvent{})
-	}
-	s.bus.publish(hardwareStopRequestEvent{})
-}
-
-type directFailure struct {
-	err error
-	stateFields
-	*broadcastContext `json: "-"`
-}
-
-func newDirectFailure(ctx *broadcastContext, err error) *directFailure {
-	return &directFailure{broadcastContext: ctx, err: err}
-}
-func (s *directFailure) enter() {
-	notifyMsg := "entering direct broadcast failure state"
-	notifyKind := broadcastGeneric
-	if s.err != nil {
-		if errEvent, ok := s.err.(errorEvent); ok {
-			notifyKind = errEvent.Kind()
-		}
-		notifyMsg = fmt.Sprintf("entering direct broadcast failure state due to: %v", s.err)
-	}
-	s.logAndNotify(notifyKind, notifyMsg)
-
-	err := s.man.StopBroadcast(context.Background(), s.cfg, s.store, s.svc)
-	if err != nil {
-		s.log("could not stop broadcast on entry: %v", err)
-	} else {
-		s.bus.publish(finishedEvent{})
-	}
-	s.bus.publish(hardwareStopRequestEvent{})
-}
-
-func (s *directFailure) MarshalJSON() ([]byte, error) {
-	return json.Marshal(struct{ Err string }{Err: s.err.Error()})
-}
-
-func (s *directFailure) UnmarshalJSON(data []byte) error {
-	aux := struct{ Err string }{}
-	if err := json.Unmarshal(data, &aux); err != nil {
-		return err
-	}
-	s.err = errors.New(aux.Err)
-	return nil
-}
-
 func updateBroadcastBasedOnState(state state, cfg *BroadcastConfig) {
 	switch state.(type) {
 	case *vidforwardPermanentLive:
